Combine single mic and virtual consultation into one unified demo
Replace separate subdirectories with a single set of files at the
AmbientScribe root. index.html has a mode toggle (radio buttons)
to switch between single-microphone diarization and virtual
consultation (doctor+patient merged streams). audio.ts exposes
getMicrophoneStream() and getRemoteParticipantStream() as two
separate methods returning a stream, plus mergeMediaStreams().
