--- Web/AmbientScribe/singleMicrophone/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Corti - Single Microphone</title>
-</head>
-<body>
-  <h1>Single Microphone - AmbientScribe</h1>
-  <button id="end-call">End Call</button>
-  <p>Open the browser console to see transcripts and facts.</p>
-  <script type="module" src="./client.ts"></script>
-</body>
-</html>
